refactor: 루트 예외를 상속 받도록 변경 및 패키지 변경
