+++ jdbc/src/main/java/org/springframework/dao/EmptyResultSetException.java
@@ -1,0 +1,5 @@
+package org.springframework.dao;
+
+public class EmptyResultSetException extends DataAccessException {
+
+}
