+++ jdbc/src/main/java/org/springframework/dao/IncorrectResultSizeDataAccessException.java
@@ -1,0 +1,8 @@
+package org.springframework.dao;
+
+public class IncorrectResultSizeDataAccessException extends DataAccessException {
+
+    public IncorrectResultSizeDataAccessException(String message) {
+        super(message);
+    }
+}
