+++ jdbc/src/main/java/org/springframework/dao/InvalidArgsException.java
@@ -1,0 +1,8 @@
+package org.springframework.dao;
+
+public class InvalidArgsException extends DataAccessException {
+
+    public InvalidArgsException(String message) {
+        super(message);
+    }
+}
