+++ jdbc/src/main/java/org/springframework/dao/ResultSetConvertException.java
@@ -1,6 +1,6 @@
-package org.springframework.jdbc.core;
+package org.springframework.dao;
 
-public class ResultSetConvertException extends RuntimeException {
+public class ResultSetConvertException extends DataAccessException {
 
     public ResultSetConvertException(String message, Throwable e) {
         super(message, e);
