--- jdbc/src/main/java/org/springframework/jdbc/core/EmptyResultSetException.java
@@ -1,5 +1,0 @@
-package org.springframework.jdbc.core;
-
-public class EmptyResultSetException extends RuntimeException {
-
-}
