--- jdbc/src/main/java/org/springframework/jdbc/core/IncorrectResultSizeDataAccessException.java
@@ -1,8 +1,0 @@
-package org.springframework.jdbc.core;
-
-public class IncorrectResultSizeDataAccessException extends RuntimeException {
-
-    public IncorrectResultSizeDataAccessException(String message) {
-        super(message);
-    }
-}
