--- jdbc/src/main/java/org/springframework/jdbc/core/InvalidArgsException.java
@@ -1,8 +1,0 @@
-package org.springframework.jdbc.core;
-
-public class InvalidArgsException extends RuntimeException {
-
-    public InvalidArgsException(String message) {
-        super(message);
-    }
-}
