+++ jdbc/src/main/java/org/springframework/jdbc/core/JdbcTemplate.java
@@ -8,6 +8,8 @@ import java.util.List;
 import javax.sql.DataSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.dao.IncorrectResultSizeDataAccessException;
+import org.springframework.dao.InvalidArgsException;
 
 public class JdbcTemplate {
