+++ jdbc/src/main/java/org/springframework/jdbc/core/ObjectConverter.java
@@ -11,6 +11,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import org.springframework.dao.EmptyResultSetException;
+import org.springframework.dao.ResultSetConvertException;
 
 public class ObjectConverter {
